test(felte): add tests for dom mutations
